+++ src/RoadCaptain.Runner/MainWindow.xaml.cs
@@ -2,9 +2,7 @@
 // Licensed under Artistic License 2.0
 // See LICENSE or https://choosealicense.com/licenses/artistic-2.0/
 
-using System.Diagnostics;
 using System.Windows;
-using System.Windows.Documents;
 using System.Windows.Input;
 using RoadCaptain.Ports;
 using RoadCaptain.Runner.ViewModels;
@@ -18,7 +16,7 @@ namespace RoadCaptain.Runner
     {
         public MainWindow(MainWindowViewModel viewModel, IGameStateReceiver gameStateReceiver)
         {
-            gameStateReceiver.Register(null, null, viewModel.UpdateGameState);
+            gameStateReceiver.Register(route => viewModel.Route = route, null, viewModel.UpdateGameState);
 
             DataContext = viewModel;